refactor: move ClientProxy to forge layer
